Split out libc tests into one enclave binary per test
Previously, the libc tests were all run as part of a single enclave
binary. Additionally, they would fail after a single test, meaning other
tests were not run after the first failure.

Add a compile-time option to allow compiling each libc test as a
separate enclave by passing -DOE_SEPARATE_LIBC_TESTS=ON to cmake.
