+++ cmake/add_enclave.cmake
@@ -165,7 +165,8 @@ macro(add_enclave_optee)
   set(CMAKE_EXE_EXPORTS_C_FLAG)
 
   string(REPLACE "gcc" "ld" LINKER ${CMAKE_C_COMPILER})
-  set(CMAKE_C_LINK_EXECUTABLE "${LINKER} <CMAKE_CXX_LINK_FLAGS> <LINK_FLAGS> <OBJECTS> <LINK_LIBRARIES> -lgcc -o <TARGET>")
+
+  set(CMAKE_C_LINK_EXECUTABLE "${LINKER} <CMAKE_C_LINK_FLAGS> <LINK_FLAGS> <OBJECTS> <LINK_LIBRARIES> -lgcc -o <TARGET>")
   set(CMAKE_CXX_LINK_EXECUTABLE "${LINKER} <CMAKE_CXX_LINK_FLAGS> <LINK_FLAGS> <OBJECTS> <LINK_LIBRARIES> -lgcc -o <TARGET>")
 
   # Generate linker script from template.
