+++ tests/libc/CMakeLists.txt
@@ -7,4 +7,13 @@ if (BUILD_ENCLAVES)
     add_subdirectory(enc)
 endif()
 
-add_enclave_test(tests/libc libc_host libc_enc)
+include(./tests.cmake)
+
+if (OE_SEPARATE_LIBC_TESTS)
+    foreach(testcase ${LIBC_TESTS})
+        get_testcase_name(${testcase} name "${PROJECT_SOURCE_DIR}")
+        add_enclave_test(tests/libc-${name} libc_host libc-${name}_enc)
+    endforeach(testcase)
+else ()
+    add_enclave_test(tests/libc-all libc_host libc_enc)
+endif ()
