+++ tests/libc/enc/run_tests_all.c.in
@@ -1,0 +1,11 @@
+// Copyright (c) Open Enclave SDK contributors.
+// Licensed under the MIT License.
+
+#include "helpers.h"
+
+int run_tests()
+{
+    int ret = 0;
+    @libc_tests@
+    return ret;
+}
