--- tests/libc/enc/tests.c.in
@@ -1,15 +1,0 @@
-// Copyright (c) Open Enclave SDK contributors.
-// Licensed under the MIT License.
-
-// This is a template used by CMake.
-
-extern int run_test(
-    const char* name,
-    int (*main)(int argc, const char* argv[]));
-
-int run_tests(void)
-{
-    int ret = 0;
-@run_test@
-    return ret;
-}
